LLM/lanchain/requirements.txt
langchain
openai
pypdf
tiktoken
pinecone-client
optimum